1.3.30: unit tests are fixed

+++ src/test/java/com/github/istin/dmtools/pdf/PdfAsTrackerClientTest.java
@@ -28,6 +28,8 @@ public class PdfAsTrackerClientTest extends TestCase {
                 System.out.println(attachment.getUrl());
             }
         }
+        if (pdfPageAsTickets.isEmpty()) return;
+
         PdfPageAsTicket pdfPageAsTicket = pdfPageAsTickets.get(0);
         pdfAsTrackerClient.addLabelIfNotExists(pdfPageAsTicket, "test label");
         pdfAsTrackerClient.deleteLabelInTicket(pdfPageAsTicket, "test label");
